+++ ai/azure/fake/src/main/kotlin/org/http4k/connect/azure/ChatCompletionGenerator.kt
@@ -23,9 +23,9 @@ fun interface ChatCompletionGenerator : (ModelCompletion) -> List<Choice> {
 val ChatCompletionGenerator.Companion.ReverseInput
     get() = ChatCompletionGenerator { req ->
         req.content().flatMap { m ->
-            m.content.mapIndexed { i, content ->
+            m.content?.mapIndexed { i, content ->
                 Choice(i, ChoiceDetail(Role.System, content.text?.reversed() ?: "", null), null, FinishReason.stop)
-            }
+            } ?: emptyList()
         }
     }
 
@@ -41,7 +41,7 @@ fun ChatCompletionGenerator.Companion.LoremIpsum(random: Random = Random(0)) = C
  */
 val ChatCompletionGenerator.Companion.Echo
     get() = ChatCompletionGenerator { req ->
-        req.choices(req.content().first { it.role == User }.content.first().text ?: "")
+        req.choices(req.content().first { it.role == User }.content?.first()?.text ?: "")
     }
 
 private fun ModelCompletion.choices(msg: String) = (if (stream) msg.split(" ").map { "$it " } else listOf(msg))